backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
django==5.1.7
djangorestframework==3.15.2
firebase-admin==6.7.0
future==1.0.0
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.18.3
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-gax==0.12.5
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
msgpack==1.1.0
oauth2client==4.1.3
ply==3.8
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-core==2.27.2
pyjwt==2.10.1
pyparsing==3.2.1
pyrebase4==4.8.0
python-dotenv==1.0.1
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rest-framework-simplejwt==0.0.2
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
